Implement parallel downloading
Add parallel downloads for bato.to and DynastyScans. The default
number of threads that are used for parallel downloads is 4.

Futures are used to implement this functionality, and the files are
assigned to the list of files in the correct position as the futures
complete, allowing the progress bar to be updated in a quicker fashion
than it would be possible if the futures were iterated in the order they
were submitted, as downloads may not complete sequentially.

For bato.to, image URLs are predicted based on the previous URL. If that
prediction fails in some way, resort to scraping the page again.

As some bato.to series apparently use a different image URL scheme, a second
regular expression is needed for those cases; both cases are tried, falling
back to scraping the page should neither work.