Implement parallel 'cum update'
Only the series_by_url call is parallelised, as it makes up the
majority of the time spent updating, and allows to be parallelised
without messing about with the database.

The download_threads config setting is used to get the number of
parallel threads to spawn.

A progress bar was added.

Warnings generated by the exceptions are now printed at the end, to
avoid interfering with the progress bar.