Explicitly close streaming requests
According to the Requests documentation in the "Body Content Workflow"
heading, the Request objects of streaming requests may not be released
back into the request pool if the content is not fully consumed and
the request is not explicitly closed.

While the page downloads do consume all of their data, we can close
those requests explicitly as soon as we finish just to be sure. One
area where a streaming request may have gone unclosed was if batoto
failed to predict an image url - the request body would go unread.