Do not write default config values back to file
By not writing default values back to the config file, we can
distinguish between user-set values and default values, which
means we can effectively change default values for existing
installations at a later date.

Attributes of the configuration which are changed after loading
is complete are counted as changes meant to be reflected in the
persistent configuration.