Change temporary page download file handling.
Before, pages would be downloaded to temporary files and then closed
after the .zip file was created. This however results in
PermissionError on Windows, where the file cannot be reopened by the
zipfile library. Thus temporary files are not set to delete on close
and rather are deleted by hand after they have been added to the .zip
file.

Should fix the PermissionError reported in #27.