Fix Specials for some DynastyScans series
Some series have their specials labelled in a certain way which
previously made cum overwrite regular chapters with the specials.

This changes the regex to have named groups, and a group to catch
whether something is a chapter or a special is introduced.

A unit test which checks one series that previously exhibited this
issue is added, in which the chapter names and the filenames are
checked to be unique.