dynastyscans: Scrape using JSON endpoints
dynasty-scans.com has largely undocumented JSON endpoints, which
this change makes use of. Instead of parsing through fragile HTML,
we can now enumerate the data we need from the JSON endpoints.

This reduces the amount of requests we need to make, particularly
for the get_groups method. It also reduces code complexity somewhat,
as we now no longer need to iterate through the HTML DOM multiple
times to get all the info we need.

For now, headings such as "Volume 1" are skipped. These could be used
for volume specific shenanigans in the future if need be.