config: check config for malformed format
If the configuration file fails to decode as JSON, we raise a new
exception called "ConfigError", which contains a string of the config
file contents, the cursor position at which the error was encountered
and a message. The cum CLI then catches this exception during config
initialisation, and outputs a beautifully formatted snippet from the
configuration file with a symbol pointing at the error location
instead of an ugly uninformative traceback.