Use a fallback for Batoto chapter info scraping
Scraping Batoto series pages sometimes confuses BeautifulSoup and/or
its HTML parser, where it will read "<img />TEXT" as "<img>TEXT</img>",
causing exceptions when the text element cannot be found as the next
sibling of the image tag but rather as the string value of the image
tag. This commit falls back on using the string value of the image tag
if next sibling returns None.

Closes #46