Add version support in MangaDex name matching
There are some instances of names containing the chapter version
information in the style of "Vol. 2 Ch. 8 v2 - Read Online", which
causes the entire string to be interpreted as the name as there is no
matching for the " v2". This change adds an optional non-capturing group
to also match names containing the version number in this manner.