scraper: add proper support for retrying pages
note that this requires an optional change to be made to
scrapers.  if you want your scraper to support retrying,
you must pass the url of the image to the worker payload
so that it can be re-requested if necessary.  if the worker
fails to download an image and the scraper has not passed
this parameter, an error message will be emitted and the
old behavior will be used, i.e., crash everything.